requirements.txt
aiohttp
aiosignal
altair
anyio
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-timeout
attrs
backcall
beautifulsoup4
bleach
blinker
cachetools
certifi
cffi
charset-normalizer
chromadb
click
cmake
colorama
comm
dataclasses-json
debugpy
decorator
defusedxml
entrypoints
executing
faiss-cpu
fastjsonschema
filelock
fqdn
frozenlist
gitdb
GitPython
greenlet
huggingface-hub
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
isoduration
jedi
Jinja2
jsonpointer
jsonschema
jupyter_client
jupyter_core
jupyter-events
jupyter_server
jupyter_server_terminals
jupyterlab-pygments
langchain
lit
markdown-it-py
markupsafe
marshmallow-enum
marshmallow
matplotlib-inline
mdurl
mistune
mpmath
multidict
mypy-extensions
nbclassic
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numpy
nvidia-cublas-cu11
nvidia-cuda-cupti-cu11
nvidia-cuda-nvrtc-cu11
nvidia-cuda-runtime-cu11
nvidia-cudnn-cu11
nvidia-cufft-cu11
nvidia-curand-cu11
nvidia-cusolver-cu11
nvidia-cusparse-cu11
nvidia-nccl-cu11
nvidia-nvtx-cu11
openai
openapi-schema-pydantic
packaging
pandas
pandocfilters
parso
pdfminer
pdfminer.six
pexpect
pickleshare
pillow
protobuf
pyarrow
pycparser
pydantic
pydeck
Pygments
pympler
pyrsistent
python-dateutil
python-dotenv
python-json-logger
pytz-deprecation-shim
pytz
pyyaml
regex
requests
rfc3339-validator
rfc3986-validator
rich
scikit-learn
scipy
sentence-transformers
sentencepiece
Send2Trash
setuptools
six
smmap
sniffio
soupsieve
SQLAlchemy
stack-data
streamlit
streamlit-chat
sympy
tenacity
terminado
threadpoolctl
tinycss2
tokenizers
toml
toolz
torch
torchvision
tornado
tqdm
transformers
triton
typing-extensions
typing-inspect
tzdata
tzlocal
urllib3
uri-template
validators
wcwidth
webcolors
webencodings
websocket-client
wheel
yarl
zipp